2023-10-09 Modificacion de CRUDcontroller para recibir peticiones AJAX

+++ app/Controllers/Reticulas/CrudController.php
@@ -47,11 +47,22 @@ class CrudController extends BaseController
     // DB operations
     public function delete()
     {
-        $id = $this->request->getPost('id');
-
-        $this->model->delete($id);
-
         dd('Registro eliminado: ' . $this->name);
+
+        try {
+            if (!$this->request->isAJAX()) {
+                throw new Exception('No se encontró el recurso', 404);
+            }
+            // $data = $this->request->getPost();
+
+            $id = $this->request->getPost('id');
+
+            $this->model->delete($id);
+
+            return $this->response->setStatusCode(200)->setJSON(['success' => true]);
+        } catch (Exception $e) {
+            return $this->response->setStatusCode($e->getCode())->setJSON(['error' => $e->getMessage()]);
+        }
     }
 
     // Model's save method performs both insert and update
